pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nora-pipeline"
version = "0.1.0"
description = "Norm-Aware Agent for Reliable ML Training"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Author", email = "author@example.com"}]
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "pyyaml>=6.0",
    "numpy>=1.21",
    "pandas>=1.3",
    "matplotlib>=3.4",
    "omegaconf>=2.1",
    "dataclasses-json>=0.5.7",
    "catboost>=1.1",
    "scikit-learn>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "isort>=5.10",
    "flake8>=4.0",
    "mypy>=0.950",
]
docker = [
    "docker>=6.0",
]

[project.urls]
Homepage = "https://github.com/alinutzal/norm-aware-agent39"
Documentation = "https://github.com/alinutzal/norm-aware-agent39/wiki"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
